# Use []float64 for histogram boundaries, not []metric.Number