--- a/sdk/metric/aggregator/histogram/benchmark_test.go
+++ b/sdk/metric/aggregator/histogram/benchmark_test.go
@@ -52,20 +52,8 @@ func benchmarkHistogramSearchFloat64(b *testing.B, size int) {
 func BenchmarkHistogramSearchFloat64_1(b *testing.B) {
 	benchmarkHistogramSearchFloat64(b, 1)
 }
-func BenchmarkHistogramSearchFloat64_2(b *testing.B) {
-	benchmarkHistogramSearchFloat64(b, 2)
-}
-func BenchmarkHistogramSearchFloat64_3(b *testing.B) {
-	benchmarkHistogramSearchFloat64(b, 3)
-}
-func BenchmarkHistogramSearchFloat64_4(b *testing.B) {
-	benchmarkHistogramSearchFloat64(b, 4)
-}
 func BenchmarkHistogramSearchFloat64_8(b *testing.B) {
 	benchmarkHistogramSearchFloat64(b, 8)
-}
-func BenchmarkHistogramSearchFloat64_12(b *testing.B) {
-	benchmarkHistogramSearchFloat64(b, 12)
 }
 func BenchmarkHistogramSearchFloat64_16(b *testing.B) {
 	benchmarkHistogramSearchFloat64(b, 16)
@@ -75,6 +63,18 @@ func BenchmarkHistogramSearchFloat64_32(b *testing.B) {
 }
 func BenchmarkHistogramSearchFloat64_64(b *testing.B) {
 	benchmarkHistogramSearchFloat64(b, 64)
+}
+func BenchmarkHistogramSearchFloat64_128(b *testing.B) {
+	benchmarkHistogramSearchFloat64(b, 128)
+}
+func BenchmarkHistogramSearchFloat64_256(b *testing.B) {
+	benchmarkHistogramSearchFloat64(b, 256)
+}
+func BenchmarkHistogramSearchFloat64_512(b *testing.B) {
+	benchmarkHistogramSearchFloat64(b, 512)
+}
+func BenchmarkHistogramSearchFloat64_1024(b *testing.B) {
+	benchmarkHistogramSearchFloat64(b, 1024)
 }
 
 func benchmarkHistogramSearchInt64(b *testing.B, size int) {
@@ -103,20 +103,8 @@ func benchmarkHistogramSearchInt64(b *testing.B, size int) {
 func BenchmarkHistogramSearchInt64_1(b *testing.B) {
 	benchmarkHistogramSearchInt64(b, 1)
 }
-func BenchmarkHistogramSearchInt64_2(b *testing.B) {
-	benchmarkHistogramSearchInt64(b, 2)
-}
-func BenchmarkHistogramSearchInt64_3(b *testing.B) {
-	benchmarkHistogramSearchInt64(b, 3)
-}
-func BenchmarkHistogramSearchInt64_4(b *testing.B) {
-	benchmarkHistogramSearchInt64(b, 4)
-}
 func BenchmarkHistogramSearchInt64_8(b *testing.B) {
 	benchmarkHistogramSearchInt64(b, 8)
-}
-func BenchmarkHistogramSearchInt64_12(b *testing.B) {
-	benchmarkHistogramSearchInt64(b, 12)
 }
 func BenchmarkHistogramSearchInt64_16(b *testing.B) {
 	benchmarkHistogramSearchInt64(b, 16)
@@ -126,4 +114,16 @@ func BenchmarkHistogramSearchInt64_32(b *testing.B) {
 }
 func BenchmarkHistogramSearchInt64_64(b *testing.B) {
 	benchmarkHistogramSearchInt64(b, 64)
+}
+func BenchmarkHistogramSearchInt64_128(b *testing.B) {
+	benchmarkHistogramSearchInt64(b, 128)
+}
+func BenchmarkHistogramSearchInt64_256(b *testing.B) {
+	benchmarkHistogramSearchInt64(b, 256)
+}
+func BenchmarkHistogramSearchInt64_512(b *testing.B) {
+	benchmarkHistogramSearchInt64(b, 512)
+}
+func BenchmarkHistogramSearchInt64_1024(b *testing.B) {
+	benchmarkHistogramSearchInt64(b, 1024)
 }