Reading through the histogram aggregator was confusing because the boundaries are specified as []metric.Number, but despite this were expected to be floating point numbers. This makes the boundaries []float, which is less confusing.

This also adds a benchmark. I considered whether binary search outperforms linear search in any practical-sized histogram, and the answer is no. Binary search beats linear search for histograms sized >= 256 to 512 elements.

Resolves #534.
Resolves #759.